Add cmake and fix some warnings

+++ evolutionary-algorithm.h
@@ -21,7 +21,7 @@ class EvolutionarySearch {
 	QuantumTestGenerator quant_test;
 	
 	void SetDnaLength (int);
-	void SetMutationRate (int);
+	void SetMutationRate (double);
 	void SetGenerationSize (int);
 	vector<int> GenerateDna ();
 	void Init ();
